backend/requirements.txt
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.7.0

pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0

python-dotenv==1.0.0
numpy>=1.26.0
tiktoken==0.5.2
